Add AuthService tests and types for login and user profile responses; remove outdated test file
