+++ types/modules/auth.ts
@@ -1,0 +1,11 @@
+export type LoginResponse = {
+  success: boolean;
+};
+
+export type UserProfile = {
+  email: string;
+};
+
+export type UserProfileResponse = {
+  user: UserProfile;
+};
